requirements.txt
google-api-python-client==2.108.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
textblob==0.17.1
loguru==0.7.2
pytest==7.4.3
requests==2.31.0
nltk==3.8.1
scikit-learn==1.3.2
